Resource Manager Publish Profile > Issue with rest url when slots used
Fix up the publisheSettingsUri to use the correct format for getting the credentials when using a slot or not.

When not using a slot, it will be ..Microsoft.Web/sites/your-site-name/config/publishingCredentials...

When you are using a slot it will be:
Microsoft.Web/sites/your-site-name/slots/your-slot-name/config/publishingCredentials...

My assumption here is that in the Octopus UI, the user will denote the use of a slot by formating the website name <websitename>/<slotname> when using a custom expression.

I think it would be good to get slot detection into the UI for that feature in Octopus